ai-workflow/requirements.txt
fastapi>=0.115.0
uvicorn>=0.27.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
openai>=1.12.0
pydantic>=2.6.0
python-dotenv>=1.1.0
PyJWT[crypto]>=2.8.0
